updated quantity to decimal point for more

+++ water_proj/water_app/models.py
@@ -4,7 +4,7 @@ from django.contrib.auth.models import User
 
 class WaterIntakeModel(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    quantity = models.PositiveIntegerField()
+    quantity =  models.DecimalField(max_digits=8, decimal_places=2)
     date = models.DateField(default=timezone.now)
     time_added = models.DateTimeField(auto_now_add=True)
